+++ attributes/sensu_client_setup.rb
@@ -1,0 +1,2 @@
+default["sensu"]["directory_mode"] = "0775"
+default["sensu"]["log_directory_mode"] = "0775"