+++ recipes/logstash_nginx_modifications.rb
@@ -1,3 +1,4 @@
 
-execute "chmod 755 /var/log/nginx"
+node.set['nginx']['log_dir_perm'] = '755'
+
 execute "chmod 644 $(find /var/log/nginx -type f)"