+++ recipes/logstash_shared_post_setup.rb
@@ -1,5 +1,5 @@
 
-instance_name = node['roles'].include?('logstash_server') ? node['elkstack']['config']['logstash']['instance_name'] : node['elkstack']['config']['logstash']['agent_name']
+instance_name = node['roles'].include?(node['TheCheftacularCookbook']['role_maps']['logstash_server']) ? node['elkstack']['config']['logstash']['instance_name'] : node['elkstack']['config']['logstash']['agent_name']
 
 nginx_template = { 'nginx' => 'logstash/nginx.erb' }
 
@@ -8,4 +8,12 @@ logstash_pattern instance_name do
   templates_cookbook  'TheCheftacularCookbook'
   owner               'logstash'
   group               'logstash'
+end
+
+if node['roles'].include?(node['TheCheftacularCookbook']['role_maps']['logstash_nginx_logging'])
+  group 'adm' do
+    action :modify
+    members 'logstash'
+    append true
+  end
 end