+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.4.4
+
+* Bump uchiwa version and fix sensu gem installation regarding that bump
+
+## 1.4.3
+
+* Minor fixes to sensu recipes
+
 ## 1.4.1
 
 * Major fixes for elasticsearch cookbook update affecting grafana cookbook